requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.2
requests==2.32.3
tomli==2.0.1
urllib3==2.2.1
structlog==24.2.0
curlify==2.2.1
retrying==1.3.4
pydantic==2.7.4
pyhamcrest==2.1.0